requirements.txt
aiohttp==3.8.5
aiolimiter==1.0.0
python-dotenv==1.0.0
